Beta diversity between consecutive time steps

Calculates sample dissimilarity between consecutive time points (t, t+i), within a group (subject, reaction chamber, or similar). The corresponding time difference is returned as well. The method operates on SummarizedExperiment objects, and the results are stored in colData.

getStepwiseDivergence(
  x,
  group = NULL,
  time_field,
  time_interval = 1,
  name_divergence = "time_divergence",
  name_timedifference = "time_difference",
  abund_values = "counts",
  FUN = vegan::vegdist,
  ...
)

getTimeDivergence(x, ...)

# S4 method for ANY
getTimeDivergence(x, ...)

Arguments

x

A SummarizedExperiment object.

group

optional; a single character value for specifying the grouping factor (name of a colData field). If given, the divergence is calculated per group. e.g. subject, chamber, group etc.).

time_field

a single character value, specifying the name of the time series field in colData.

time_interval

integer value indicating the increment between time steps (default: 1). If you need to take every second, every third, or so, time step only, then increase this accordingly.

name_divergence

a column vector showing beta diversity between samples over n time intervals (default: name_divergence = "time_divergence")

name_timedifference

field name for adding the time difference between samples used to calculate beta diversity (default: name_timedifference = "time_difference")

abund_values

character indicating which assay values are used in the dissimilarity estimation (default: abund_values = "counts")

FUN

a function for dissimilarity calculation. The function must expect the input matrix as its first argument. With rows as samples and columns as features. By default, FUN is vegan::vegdist.

...

Arguments to be passed

Value

a SummarizedExperimentor TreeSummarizedExperimentcontaining the sample dissimilarity and corresponding time difference between samples (across n time steps), within each level of the grouping factor.

Examples

#library(miaTime)
library(TreeSummarizedExperiment)

data(hitchip1006)
tse <- mia::transformSamples(hitchip1006, method = "relabundance")

# Subset to speed up example
tse <- tse[, colData(tse)$subject %in% c("900", "934", "843", "875")]

tse2 <- getStepwiseDivergence(tse, group = "subject",
                              time_interval = 1,
                              time_field = "time",
                              abund_values="relabundance",
                              FUN = vegan::vegdist)