Beta diversity calculation within individuals in the interval of n time steps

The dissimilarity (beta diversity) is calculated in time steps of n, between the samples of subject (individuals).Time interval can be 1 or greater than 1. The time difference between n points is also calculated. Given input, SummarizedExperiment or TreeSummarizedExperiment object, returns with the calculations inserted in columns of colData field.

check_pairwise_dist(
  x,
  distfun,
  time_interval,
  new_field = "time_divergence",
  new_field2 = "time_difference",
  time_field,
  abund_values
)

getTimeDivergence(
  se,
  field,
  time_field,
  time_interval,
  new_field = "time_divergence",
  new_field2 = "time_difference",
  abund_values = "counts"
)

Arguments

x

a SummarizedExperiment object or a TreeSummarizedExperiment of an individual with multiple samples

distfun

a function that uses vegan::vegdist as the default value to calculate beta diversity. (default: distfun = vegan::vegdist)

time_interval

integer value indicating the increment between the time steps. It can be 1 or higher than 1. (default: time_interval = 1)

new_field

a column vector showing beta diversity between samples over n time intervals (default: new_field = "time_divergence")

new_field2

a column vector showing the time difference between samples used to calculate beta diversity (default: new_field2 = "time_difference")

time_field

a single character value for specifying time series vector given in colData field.

abund_values

character indicating which abundance values are used in the assay of SE and TSE (default: abund_values = "counts")

se

A SummarizedExperiment object or TreeSummarizedExperiment

field

a single character value for specifying which grouping factor chosen from columns of colData field.

Value

a SummarizedExperimentor TreeSummarizedExperimentcontaining the beta diversity and time difference between samples in n time steps

Examples

library(miaTime)
data(hitchip1006)

sub_hitchip <- hitchip1006[,800:1151]
#> Loading required package: TreeSummarizedExperiment
#> Loading required package: SingleCellExperiment
#> Loading required package: SummarizedExperiment
#> Loading required package: MatrixGenerics
#> Loading required package: matrixStats
#> 
#> Attaching package: ‘MatrixGenerics’
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colAlls, colAnyNAs, colAnys, colAvgsPerRowSet, colCollapse,
#>     colCounts, colCummaxs, colCummins, colCumprods, colCumsums,
#>     colDiffs, colIQRDiffs, colIQRs, colLogSumExps, colMadDiffs,
#>     colMads, colMaxs, colMeans2, colMedians, colMins, colOrderStats,
#>     colProds, colQuantiles, colRanges, colRanks, colSdDiffs, colSds,
#>     colSums2, colTabulates, colVarDiffs, colVars, colWeightedMads,
#>     colWeightedMeans, colWeightedMedians, colWeightedSds,
#>     colWeightedVars, rowAlls, rowAnyNAs, rowAnys, rowAvgsPerColSet,
#>     rowCollapse, rowCounts, rowCummaxs, rowCummins, rowCumprods,
#>     rowCumsums, rowDiffs, rowIQRDiffs, rowIQRs, rowLogSumExps,
#>     rowMadDiffs, rowMads, rowMaxs, rowMeans2, rowMedians, rowMins,
#>     rowOrderStats, rowProds, rowQuantiles, rowRanges, rowRanks,
#>     rowSdDiffs, rowSds, rowSums2, rowTabulates, rowVarDiffs, rowVars,
#>     rowWeightedMads, rowWeightedMeans, rowWeightedMedians,
#>     rowWeightedSds, rowWeightedVars
#> Loading required package: GenomicRanges
#> Loading required package: stats4
#> Loading required package: BiocGenerics
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, intersect, is.unsorted,
#>     lapply, mapply, match, mget, order, paste, pmax, pmax.int, pmin,
#>     pmin.int, rank, rbind, rownames, sapply, setdiff, sort, table,
#>     tapply, union, unique, unsplit, which.max, which.min
#> Loading required package: S4Vectors
#> 
#> Attaching package: ‘S4Vectors’
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
#> Loading required package: IRanges
#> Loading required package: GenomeInfoDb
#> Loading required package: Biobase
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.
#> 
#> Attaching package: ‘Biobase’
#> The following object is masked from ‘package:MatrixGenerics’:
#> 
#>     rowMedians
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     anyMissing, rowMedians
#> Loading required package: Biostrings
#> Loading required package: XVector
#> 
#> Attaching package: ‘Biostrings’
#> The following object is masked from ‘package:base’:
#> 
#>     strsplit
hitchipTime_subset <- getTimeDivergence(sub_hitchip, field = "subject",
                                   time_interval = 1,
                                   time_field = "time")
#> Registered S3 method overwritten by 'vegan':
#>   method         from     
#>   reorder.hclust seriation